+++ lib/rundoc/code_command/rundoc_command.rb
@@ -26,7 +26,9 @@ module ::Rundoc
 
       def call(env = {})
         io.puts "Running: #{contents}"
-        eval(contents, @binding) # rubocop:disable Security/Eval
+        Rundoc.capture_stdout_stderr(io) do
+          eval(contents, @binding) # rubocop:disable Security/Eval
+        end
         ""
       end
     end
@@ -37,3 +39,4 @@ Rundoc.register_code_command(keyword: :rundoc, args_klass: Rundoc::CodeCommand::
 Rundoc.register_code_command(keyword: :"rundoc.configure", args_klass: Rundoc::CodeCommand::RundocCommandArgs, runner_klass: Rundoc::CodeCommand::RundocCommandRunner)
 
 require "rundoc/code_command/rundoc/require"
+require "rundoc/code_command/rundoc/ensure_later"